Add refresh token to authorisation flow

+++ frontend/src/pages/singleJob/index.tsx
@@ -1,45 +1,43 @@
-import { useEffect, useState } from 'react';
 import { isAxiosError } from 'axios';
 import { useNavigate } from 'react-router-dom';
 import { Job } from '../../types';
 import { deleteJob } from '../../services/jobs';
 import { useStateValue, removeJob } from '../../state';
-import Custom404 from '../custom404';
+import useErrorHandler from '../../hooks/useErrorHandler';
+import Error from '../../components/Error';
 import JobInfo from '../../components/JobInfo';
 
-const SingleJob = ({ job }: { job: Job | null | undefined }) => {
+const SingleJob = ({ job }: { job: Job }) => {
   const navigate = useNavigate();
   const [{ user }, dispatch] = useStateValue();
-  const [error, setError] = useState('');
-
-  useEffect(() => {
-    if (!user) {
-      navigate('/');
-    }
-  }, []);
+  const [error, handleError] = useErrorHandler();
 
   if (!user) {
     return null;
   }
 
-  if (!job) {
-    return <Custom404 />;
-  }
   const handleDelete = async () => {
     try {
-      await deleteJob(user.token, job.id);
+      const userId = window.localStorage.getItem('id');
+
+      if (!userId) {
+        navigate('/');
+        return null;
+      }
+
+      await deleteJob(user.token, Number(userId), job.id);
       dispatch(removeJob(job.id));
       navigate('/jobs');
     } catch (err) {
       if (isAxiosError(err)) {
-        setError(err.response?.data);
+        handleError(err.response?.data);
       }
     }
   };
 
   return (
     <main>
-      {error && <p>{error}</p>}
+      {error && <Error err={error} />}
       <JobInfo
         job={job}
         handleDelete={handleDelete}
